Complete the exercise as instructed
I have completed the exercise and compared my code with that of the solution. I noticed that the authors have used combination of flex and block layouts. I also did that without noticing. I tend to divide and plan the whole layout in terms of boxes. This gives me extra, empty boxes. For example, in index.html file, I had added an icon-wrapper which I removed after going through solution. Similarly, I added a bodytext-wrapper, which I even didn't convert in flexbox. And still kept it in the code. I have grasped the basic idea of how to align the containers and play with the axis to get a desired layout. Looking at the solutions, it feels that there are more than one ways to reach a desired layout. The only point is knowing why we are doing what we are doing. I weeded out couple of code from my CSS after reading the solution.