Complete the exercise successfully
This exercise seemed easier than before. Couple of learnings to remember: flex containers have effect only on their direct children. I wasted some time on tinkering with the rleft-right-links class to make the links unstack. Later on, realized the property of flex containers and converted all <ul> types to flex and voila! Done!

Another thing to remember, removing the margin and padding using the universal selector at the beginning of the CSS. My browser was injecting a 16px of margin in a weird way.

I haven't converted the .content class in flex container as opposed to the provided solution. I tested by resizing, everything seems working just fine. So I have kept it as it is.