Complete exercise without hiccup
This exercise was more straighforward than the modal one. The instructions were clear and not much of html work needed to be done. I have taken liberties such as not making .header a flex container because it works just fine without that.

There was an additional margin injected by the browser and after struggling for a few minutes to align the text in sidebar to the left, I used the universal selector to remove the margin and padding from all. This immediately solved the problem of injected 16px margin.